requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0
django-js-asset==2.0.0
django-mptt==0.13.4
psycopg2==2.9.3
sqlparse==0.4.2
